made mapActivity and separated driverMainPage
