+++ app/src/main/java/ca/ualberta/boost/DriverMainPage.java
@@ -6,6 +6,7 @@ import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentActivity;
 
 import android.Manifest;
+import android.app.Activity;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.location.Location;
@@ -32,15 +33,12 @@ import com.google.firebase.auth.FirebaseAuth;
  * the respective activity is launched
  */
 
-public class DriverMainPage extends FragmentActivity implements OnMapReadyCallback {
+public class DriverMainPage extends MapActivity {
 
     private static final String TAG = "DriverMainPage";
-    private static final int LOCATION_PERMISSION_REQUEST_CODE = 1234;
-    private static final int DEFAULT_ZOOM = 16;
 
-    private Boolean mLocationPermissionsGranted = false;
     private GoogleMap mMap;
-    private FusedLocationProviderClient mFusedLocationProviderClient;
+
     private Button viewRequestsButton;
     private Button logoutButton;
     private Button viewProfileButton;
@@ -51,15 +49,30 @@ public class DriverMainPage extends FragmentActivity implements OnMapReadyCallba
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_driver_main_page);
-        getLocationPermission();
 
         auth = FirebaseAuth.getInstance();
-
         viewProfileButton = findViewById(R.id.viewProfileButton);
-
         viewRequestsButton = findViewById(R.id.viewRequestsButton);
         logoutButton = findViewById(R.id.logoutButton);
+
+    }
+
+    @Override
+    protected MapFragment getMapFragment() {
+        return (MapFragment) getFragmentManager().findFragmentById(R.id.mapFragment);
+
+    }
+
+    @Override
+    protected int getLayoutResourceId() {
+        return R.layout.activity_driver_main_page;
+    }
+
+    @Override
+    protected void init() {
+        // get map from superclass
+        mMap = getMap();
+
         viewRequestsButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -81,108 +94,6 @@ public class DriverMainPage extends FragmentActivity implements OnMapReadyCallba
             }
         });
 
-
-
-    }
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-
-        if(mLocationPermissionsGranted){
-            getDeviceLocation();
-            // makes a blue dot on the map showing current location
-            mMap.setMyLocationEnabled(true);
-            // get rid of top right corner button to center location
-            mMap.getUiSettings().setMyLocationButtonEnabled(false);
-            // enable all zoom, rotate, tilt, etc gesture
-            // mMap.getUiSettings().setAllGesturesEnabled(true);
-
-
-
-        }
-    }
-
-    private void initMap() {
-        MapFragment mapFragment = (MapFragment) getFragmentManager().findFragmentById(R.id.mapFragment);
-        mapFragment.getMapAsync(DriverMainPage.this);
-    }
-
-    private void getDeviceLocation(){
-        /* get the device's current location */
-        mFusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
-
-        try{
-            if(mLocationPermissionsGranted){
-                Task location = mFusedLocationProviderClient.getLastLocation();
-                location.addOnCompleteListener(new OnCompleteListener() {
-                    @Override
-                    public void onComplete(@NonNull Task task) {
-                        if (task.isSuccessful()){
-                            Location currentLocation = (Location) task.getResult();
-                            moveCamera(new LatLng(currentLocation.getLatitude(),
-                                    currentLocation.getLongitude()), DEFAULT_ZOOM);
-                        }else{
-                            Toast.makeText(DriverMainPage.this, "unable to get current location", Toast.LENGTH_SHORT).show();
-                        }
-                    }
-                });
-            }
-        }catch (SecurityException e){
-            Log.e(TAG, "getDeviceLocation: SecurityException: " + e.getMessage());
-        }
-    }
-
-    private void moveCamera(LatLng latLng, float zoom){
-        /* moves the camera to latitude and longitude at zoom level */
-        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, zoom));
-    }
-
-    private void getLocationPermission() {
-        /*
-         * Request location permission, so that we can get the location of the
-         * device. The result of the permission request is handled by a callback,
-         * onRequestPermissionsResult.
-         */
-        String[] permissions = {Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.ACCESS_COARSE_LOCATION};
-
-        if (ContextCompat.checkSelfPermission(this.getApplicationContext(),
-                Manifest.permission.ACCESS_FINE_LOCATION)
-                == PackageManager.PERMISSION_GRANTED) {
-            if (ContextCompat.checkSelfPermission(this.getApplicationContext(),
-                    Manifest.permission.ACCESS_COARSE_LOCATION)
-                    == PackageManager.PERMISSION_GRANTED){
-                mLocationPermissionsGranted = true;
-                initMap();
-            }else{
-                ActivityCompat.requestPermissions(this, permissions, LOCATION_PERMISSION_REQUEST_CODE);
-            }
-        }else{
-            ActivityCompat.requestPermissions(this, permissions, LOCATION_PERMISSION_REQUEST_CODE);
-        }
-    }
-
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        /* look for request permission result */
-        mLocationPermissionsGranted = false;
-        switch(requestCode){
-            case LOCATION_PERMISSION_REQUEST_CODE:{
-                if(grantResults.length > 0 ){
-                    for (int grantResult : grantResults) {
-                        if (grantResult != PackageManager.PERMISSION_GRANTED) {
-                            mLocationPermissionsGranted = false;
-                            return;
-                        }
-                    }
-                    mLocationPermissionsGranted = true;
-                    // now we can initialize the map
-                    initMap();
-                }
-            }
-        }
     }
 
     //function to launch the ViewRideRequests Activity
